Issue426: MPI version of the diffusion operator
This commit reverts the implicit subroutines to original versions and adds a check at the beginning of the diff_setup
if implicit version is called using MPI. It aborts if mpi_nprocs > 1.

The implicit calculation loops over all values of lonIndex and cannot be done just for a subset of longitude values.
Instead, we would have to first do a transposition to an mpi decomposition that distributes
the 3D grid over all mpi tasks by vertical level and latitude for the application of the implicit operator in the longitude direction.
Then we have to do a transposition to have the distribution by vertical level and longitude
for the application of the operator in the latitude direction.
Finally, we need to transpose back to latitude-longitude tiles.