Issue #451: Fix 2 bugs
Include additional check on the dimensions of the computational grid
when deciding if the ensemble members need to be "padded" in the
horizontal. This is to avoid the padding for programs that do not
require it.

Also, initialize to zero the statevector used for reading the trials
in EnVar to avoid an "invalid floating point" error.