Issue #456 Analysis increment extrapolated value set to zero
In the allocation of the state vector for the analysis increment,
the extrapolation option is added and set to a zero value.
This override the default behaviour, which is to use the maximum
value of the field. This change has no impact on the results of
the tests, except for the SST test.

The other change is a call to the new GLtoLG function, which
is the main purpose of the issue. The GLtoLG function creates
the continuous ice concentration field LG.