Issue #273: Fix computing dlat/dlon for U grid.
For U-grid, it happens that for a pair of neighboring grid points,
one is on Yin and the other is on Yan. The computed max grid spacing
exceeds 1000 km threshold in this case.

The solution is to restrict the calculation of max grid spacing to
the Yin grid (lower half) of the lat2d_4/lon2d_4 arrays, defined as
lat2d_4(1:ni,1:nj/2), lon2d_4(1:ni,1:nj/2).