Issue #237: Using kdtree to construct the footprint for horizontal interp.
The original footprint operator was for LAM grids. For a global
U-grid, if part of the TOVS footprint is outside the grid domain
obs is rejected. This should now happen in the global domain.

The solution is to use kdtree to sort grid points from the two
sub-grids based on the distance from the obs, and fill the
interInfo structure with grid points falling in the footprint
radius.

Proper deallocation of the tree object with kdtree2_destroy prevents
the steady increase in memory usage.