Issue #460: Fix abort during BGCK GPSRO run.
Abort was caused by a bit overflow while writing GPSRO BURPs (trying to
write a number that was too big), itself caused by a bad extrapolation
in the GPSRO operator. All GPSRO data are interpolated, except a few
observations (less than 1 in 10000), due to a mismatch between the real
and the model topography. Some observations are below model topography.
The previous extrapolation used the lowest two levels, assuming a
linear-log profile. This may be unstable if there is a strong surface
gradient. It was substituted by a linear-log extrapolation with a
constant exp gradient (scale height 6500m), which is valid for the few
meters that have to be extrapolated and is much more robust. Only less
than 1 in 100 profiles need this extrapolation, and less than 1 in 100
observations within these profiles is extrapolated. Only a small
fraction of these profiles leads to abort (unknown, estimated of
the order of 1 GPSRO obs per 100 million). All non-extrapolated GPSRO
observations keep bit-compatibility. A typical run will contain 1-2
non-bitcompatible GPSRO observations (within 2e5 GPSRO obs), thus an
updated reference was created. The small number of observations affected
indicates that there is no qualitative impact to be expected.