Issue #237: First attempt to make kdtree thread-safe.
In kdtree2_mod, the search results (sr) is a module variable and not
thread-safe. The idea is to be able to allocate sr with total number
of threads and each thread access sr based on thread index. Since
kdtree2_create is moved outside openMP loop, the omp_get_thread_num
should show return thread index equal to zero.