Issue #237: First attempt to reduce the kdtree timing.
The maximum size of the search vector is reduced to 100. Also the
tree is not sorted to save time.

Additional timing table entries for each openMP thread calling
kdtree2_r_nearest and s2c_setupFootprintInterp are created to diagnose
the part of the code taking more time.

Fixing the getTovsFootprintRadius output foorprint value to real(4).