compile_all_plat: Avoid compiling 'splitobs' on both platforms at the same time
Since the program 'splitobs' was compiled in the same 'compile_job' as
other MIDAS programs, it happened that 'splitobs' was compiling at
exactly the same time on both platform.  As this program is compiled
directly in the code directory then it is not protected for such
behaviour.

Now, we make sure the compilation is done sequentially for 'splitobs'.