Issue #273: Reintroduce parameter 'sqlExtraDat' in all the namelists.
By adding 'order by id_obs' to all requests in 'sqlr_readSqlite', the
test '/Tests/var/EnVar/96mem_ppp_sqlite' is no longer assimilating the
SC observations despite it reads the data in the input file.  So we
prefer, remove that for now and put again those lines in each of the
namelists.

Revert "Issue #273: Remove parameter 'sqlExtraDat' in all the namelists."

This reverts commit 4014bd08a1baac8a54cf65771a012cc7c6b1865e.