Issue #237: First set of changes after code review.
Reduce number of times maxGridSpacing is printed to the listing.

Print a warning to the listing if the number of grid points found
within footprint radius is less than minNumLocalGridptsSearch.

Count the number of TOVS using footprint operator on one level and
print to the listing.

Fixing the indentation for some part of the code.

Using 'codtyp_get_codtyp' function to retrieve the codtyp of each
instrument instead of directly hardcoding the codtyp values for
getting FOV angular diameter.