Issue #470: Remove hardcoded table and sql names in code
Now sql table and column names only appear at the top of
the module in the matching table (for column names) and
other character strings for table names. This allows for
generalization in the future for obsDB format.

Also removed references to "data" table, since this name
is not used in obsDB. Instead only "header" and "body"
are used in variable names.