Issue #475: Remove tree from interpInfo structure.
Module variable tree has two versions: nl and tl/ad.

Based on the inputStateVectorType, either tree_nl or tree_tlad is
initialized in s2c_setupInterpInfo, if the tree is needed.

tree_nl/tree_tlad are only deallocated on the last call to
s2c_setupInterpInfo for the corresponding inputStateVectorType.