Issue #474: Variable transform of trials when writing to rehm.
Instead of performing the variable transform on the trials within
'inc_computeHighResAnalysis' and pass the stateVectorTrialTransformed
to 'inc_writeIncrementHighRes', the variable transform on the trials
is performed within inc_writeIncrementHighRes to get the transformed
trials, subtract it from high-res analysis to get the high-res
increments for writing to rehm file.

There is a need to pass low-res increment file to
'inc_writeIncrementHighRes' just to extract a subset of vco levels
from the trials, if needed.