Issue #474: stateVectorTrialHU is module variable in minimization_mod.
To avoid modifying the subroutines using 'simvar' as argument
(e.g. n1qn3) and change to 'simvar' interface itself,
stateVectorTrialHU_ptr is defined as pointer module variable in
minimization_mod and passed to B matrix subroutines in 'simvar',
when associated.

stateVectorTrialHU on analysis grid is only initialized when HU is
available in the trials.