Issue #474: Using 'stateVectorTrialNoZorP' for reading the trials.
The original statevector allocated for reading the trials has Z_T,
Z_M, P_T and P_M variables allocated. The trial fields do not have
these fields and there is unnecessary 4GB of memory allocated because
of these variables. A statevector with no Z/P is allocated and used
to read in the trials instead to save the memory usage.

Running on eccc-ppp4 temporarily.