Issue #474: Changes after final code review.
Name change for variables used as input argument to subroutines
inc_computeHighResAnalysis and inc_writeIncrementHighRes.

readNameList is changed to be a private subroutine available to all
the subroutine within increment_mod. The mechanism is added to return
if the namelist is already read in. The call to the original
inc_readNAMINC in the var.f90 program is removed.