Issue #474: stateVectorTrial* is stateVectorUpdate* in var.
stateVectorTrialHighRes is replaced by stateVectorUpdateHighRes and
it should be deallocated outside outer-loop. The deallocation within
inn_setupColumnsOnTrialLev is removed.

When needed stateVectorTrial is replaced by stateVectorUpdate, which is
the stateVectorUpdateHighRes on tim_nstepobsinc. stateVectorTrial is
needed for writing the final increment and should to be preserved.

The trlColumnOnAnlLev and trlColumnOnTrlLev column names are changed
to columnOnAnlLev and columnOnTrlLev.