Issue #474: Use gsv_tInterpolate when computing high-res analysis.
gsv_tInterpolate is used to obtain the high temporal resolution
analysis, stateVectorAnalHighRes, at the end of
'inc_computeHighResAnalysis' routine. The 'stateVectorAnalHighRes'
is passed as argument to 'inc_writeIncrementHighRes' routine, and
the numStep is degraded to prepare for writing the rehm* and anlm*
files. The 'stateVectorAnalHighRes' at end of each outer-loop
iteration is needed for computing the innovations during the next
iteration.

The 'HighRes' variable extension in var program indicates the
variable has the trials' high temporal and spatial resolution.

The 'run' MPI topology of EnVar/96mem_ppp is changed from 8x8x2 to
8x8x4 to prevent abort.