Issue #474: Use gsv_tInterpolate to compute high-res P0 analysis.
gsv_tInterpolate is used to obtain the high temporal resolution
P0 analysis, stateVectorPsfcHighRes, at the end of
'inc_computeHighResAnalysis' routine. The 'stateVectorPsfcHighRes'
is passed as argument to 'inc_writeIncrementHighRes' routine, and
the numStep is degraded to prepare for writing the rehm* and anlm*
files. The surface height is copied from low-res trials into
low temporal resolution P0 analysis statevector, 'stateVectorPsfc',
before writing to the file.