Issue #474: Re-read trials before writing increments to rehm.
stateVectorTrial with degraded timesteps is needed for writing the
increments to rehm file. Re-reading high resolution trials and
degrading the timesteps at the end removes the need to preserve the
stateVectorTrial when reading the trials at the beginning of the var
program. This should help to keep the memory usage low.