Issue #474: Module variable indicates next outer-loop in minimization_mod.
The module variable 'initializeForOuterLoop' is initialized to true
when 'min_minimize' is called and it is set to false at the end of
'simvar'. This module variable is passed as input argument to
oop_Htl/oop_Had to dictate the tl/ad obs operator routines execute
as if they are called for the first time in the first inner-loop
execution of every outer-loop iteration.

Deallocate the non-diagonal R matrix outside outer-loop.

To allocate some module variables in 'minimization_mod' and
'obsoperator_mod', 'utl_reallocate' routine should be used to allow
the allocation of the variable in the next outer-loop iteration.