Issue #474: Replace 'dg_vbar' with 'controlVectorIncrSum_ptr'.
In 'minimization_mod', 'dg_vbar' is sum of control vector increment
from previous outer-loops and is used for writing the hessian. It is
initialized to zero for each call to 'quasiNewtonMinimization'. In
case of multiple outer-loops, it should have the most updated value,
i.e. 'controlVectorIncrSum_ptr'.

Only nullify the 'controlVectorIncrSum_ptr' if:
- not writing the hessian, OR
- if writing the hessian, after the write is finished.