Issue #474: Adjusting zeps0 and zpes1 in quasiNewtonMinimization.
zeps0 is the desired reduction of amplitude in cost function. zeps1 is
the achieved partial reduction of cost function after a single call
to n1qn3. Changes are made so zeps1 is the value before a call to n1qn3
and zeps0 is the value after the call. zeps0 is the final reduction
amplitude and is written to the Hessian file.

The zeps0 value is propagated between outer loop iterations and is not
reset.