Issue #474: Warm start of n1qn3 when outerLoopIndex>1.
During only the first outer-loop iteration it is the presence of
the precon file that determines if the first set of iterations
(either with or without varQC depending on if varQC is activated)
are warm or cold started. For all other outer-loop iterations, they
should be warm started, since they start from the Hessian of the
previous set of iterations.
~