Issue #495:Abort in var for UMOS application
For some reasons, changes in increment_mod were compatible with the 2D_ppp unitTest but not with the UMOS (2d) application.

Changes needed to be made to pass the info to not call vco_ensureCompatibleTops when only a subset of levels from the trials need to be used when generating the analysis file.