Issue #503: Users can now set the horizontal interpolation type in bMatrixEnsemble_mod
A new namalist variable, hInterpolationDegree, was added to bMatrixEnsemble_mod to allow the users to specify the desired horizontal interpolation type for the ensembles.

This change was made to be able to get the same results when interpolating the ensembles within that module instead of using pre-interpolated ensembles generated by the ensPostProcess program.