Issue #505: remove the dependency to bgcov in bmatrixchem_mod
For some reasons, the vertical levels in bmatrixchem_mod were defined using the Bnmc file for meteorology. This creates a problem when no Bnmc file for meteorology is provided (i.e. when only the ensembles are used to define B). From now on, the code rely on ./bgchemcov to defined the vertical levels.