Issue #474: Removing stateVectorAnalHighRes.
stateVectorUpdateHighRes is used to read trials. At the end of each
outer-loop, the high-res analysis is computed and written directly
to stateVectorUpdateHighRes. There is no need for temporary
stateVectorAnalHighRes to store the high-res analysis and write back
to stateVectorUpdateHighRes.

Outside outer-loop, stateVectorUpdateHighRes is also used to compute
the low-res analysis and increments for writing to the file.