Issue #474: gvt_computeStateVectorHeight called before variable transform.
The retrieved stateVector height from the stateVectorUpdateHighRes
should be on the analysis grid before getting passed to variable
transform routines. gvt_computeStateVectorHeight is called to do the
interpolation before the stateVector height is passed to
'gsv_transform'.

Return the 'gvt_computeStateVectorHeight' routine if the stateVector
height is already initialized (heightTrialsInitialized=.true.)