Issue #509: Added option to impose a minimum value on HU in Bens with HU as CV
When using LQ as the control variable in Bens, it is possible to impose a minimum value on HU from the ensemble of trials. However, when using HU as the control variable, no minimum value could be impose which prevent a proper comparison between the two control variables.

The variable "huMinValue" in bMatrixEnsemble_mod is now applicable to HU when HU is the control variable. When thisvariable is missing when HU is the humidity control variable, no minimum value is applied.