Issue #519: Some cleanup in innovation_mod
The comments and the code related to height corrections in inn_setupBackgroundColumnsAnl were modified and improved. The code was still mentionning that heights were computed here whereas this was not the case anymore. The heights are now simpliy interpolated with the other 3D variables in the above call to col_vintprof.

Other technical changes were also made throughout this module to make it more aligned with our standards.

These changes have no impact on the results.