Issue #519: Clarification of the surface obs related part in obsoperators_mod
In oop_sfc_nl and its TL/AD versions, it was not clear what the height of the surface observations (zhhh) really meant. After digging into the code, it was found that was the height of the obs validity and not (always) the height of the stations. More user-friendly variable names are now used. A relatively important cosmetic upgrade of the code in this module was also performed, mainly for surface obs. Nevertheless, a lot of work remains to bring this module to our top standards.