Issue #474: Removing redundant check to return in gvt_setupRefFromStateVector.
The logic of returning when the stateVectorRef contains nonZero
values is not needed in gvt_setupRefFromStateVector. For the outer
loop we call gvt_setupRefFromStateVector once per outer loop iteration.
Previously we were calling it for every iteration, so we needed a way
of knowing if the stateVectorRef is already initialized.