Issue #391: Set header flag bit 15 on to indicate OBS_BCOR is added to AI data.
For AI data, we set the header flag bit 15 to indicate that the bias correction
is ADDED to the raw observation to get the corrected value as opposed to the old
way where the bias correction was SUBTRACTED. If flag bit 15 is off, then existing
corrections are reversed by adding OBS_BCOR to the corrected value. Otherwise,
OBS_BCOR is subtracted from the corrected value to reverse the correction.