new integer8 arrays for id_obs and id_data
To allow random IDs, id_data and id_obs are now read in 64-bit
integers arrays outside of matdata.

The logic for reading sqlite files was changed to iterate over body
entries and write headers at the end of id_data batches.

This change demands that the order in which sqlite files are read to be
enforced through a hard coded "order by" sqlite queries.

Minor modifications to the wind conversion routine were necessary
to prevent it from excluding data when wind speed appeared before
direction.

Aeolus tests proved sensitive to the order in which data is read.
A bug is corrected and namelists for these tests have been modified
with the advice of Stéphane Laroche.

In this version, the sqlExtraDat config key is deprecated.
If this key is present, its content is ignored and a warning is printed.