Issue #541: Prescribe numIterMaxInnerLoop in NAMVAR.
NITERMAX from NAMMIN is reverted to scalar. numIterMaxInnerLoop is a
new namelist variable in NAMVAR.

numIterMaxInnerLoop is used to pass number of inner-loop iterations
from var to min_minimize only when there is outer loop. Otherwise,
the scalar NAMMIN namelist variable NITERMAX is used. The program
aborts if NITERMAX is specified when there is an outer loop.