Issue #557: Added scaling option for the LAM Bnmc stdDev
The capability to scale (taper) the LAM Bnmc stdDev was added in calcstatslam_mod. Previously, an external program was used for that purpose.

At the same time, a small bug was corrected in diagBmatrix. The initialization of the HU reference state for the LQ -> HU conversion was missing. This appears to be a leftover of a previous redesign of the variable transmform module.