Issue #569  Know the type of profile for GPS-RO
A structure has been created (gps_vRO_iVarCode) to know if the profile for GPS-RO corresponds
to a burf code 15036, this is a refractivity profile or if it is 15037, this is a bending profile.
The objective in this first issue (to apply bending/refractivity) is to know what type of profile
is treated, independent of the nml.