Issue #573: Rename 'primary key' for 'split on key'.
In the SQL world, "primary key" is a very special word with a very
special meaning but in the context here, by "primary key", we were
meaning the column which is used to filter the data.  So we rename
'primary key' to 'split-on-key' which is more representative of what
it represents.