Issue #563: New call for the Days Since Last Obs calculation
The DSLO is updated by the subroutine aer_daysSinceLastObs,
given the newly assimilated observations.
Also a bug fix was applied. The subroutines filt_iceConcentration and
filt_backScatAnisIce are now called by the program. The subroutine
filt_backScatAnisIce rejects the ASCAT observations where the wind speed
is lower than 4 m/s. So that changes the results of the analysis-error
calculation with respect to the previous version of the code.