Issue #559: Loading environment variables to setup MPI according to 'inteloneapi-2022.1.2'
This is much faster (less than 0.2 seconds) than loading the compiler
(more than 3 seconds) and so will not need to set the compiler
throught he calling script.  This will be hard-coded as it is unlikely
that it will change often.

Unfortunately, this changes the results of two tests: '/Tests/var/vbc'
and '/Tests/genCoeff' by a very small amount.