Issue #604: Fix mpi reduce operations in `biasCorrectionSat_mod`
Replace standard rpn_comm subroutines with our own versions
that ensure the order of summation for floating points values
is always the same. Without this change, the results obtained
on robert and underhill were slightly different for some
system tests.

This affects the results of 'genCoeff' and 'var/vbc' by a very small
margin.