Issue #609: s/r enkf_getModulatedMember to modulate ens member.
The new enkf_getModulatedMember subroutine is used to compute
vertical localization matrix, and the corresponding eigenvectors/
eigenvalues, and the modulated member. The modulated member is used
in the NL operator to compute H(x_b).

numRetainedEigen is new NAMLETKF namelist variable and is the number
of retained eigenVectors that will be used to compute the modulated
the ensemble members.