Issue #609: Get modulated ensemble member forecast.
The calculation of modulated member in enkf_getModulatedMember is
complete now. The modulated member perturbation is equal to columns of
eigenvalue^0.5*eigenvector*backgroundEnsemblePert times scalar
(nEns * numRetainedEigen / (nEns - 1)) ** 0.5, according to Eq.2 and 6
of Bishop et al. (2017).

The computed modulated member perturbation is added to ensemble mean
to result in the modulated ensemble member forecast.