Issue #609: Use stateVector4D without Z/P in letkf to compute modulated member.
For computing innovations in letkf, the ensemble member is copied to
stateVector4D that does not have 3D height/pressure. statevVector4D
is copied to stateVectorWithZandP4D afterwards. For computing the
modulated member, statevVector4D should be used since ensemble mean
does not have 3D height/pressure (the perturbation for height/pressure
can not be computed). The resulting modulated state stateVector4Dmod
(from enkf_getModulatedState) is copied to stateVectorWithZandP4D which
is used in NL operator. Surface height needs to be copied to
stateVectorWithZandP4D before calling the NL operator.

In enkf_getModulatedState, the eigenVector levIndex used for 2D
variables should be nLev. Other changes to this subroutine
include:
  - The tolerence for computing engenValues/Vectors is reduced to
    1.0D-50.
  - Adding some write statements in letkf and enkf_getModulatedState
    for debugging.
  - Turn off openMP loop in enkf_getModulatedState.