Issue #569: Process GPSRO profiles based on data
Standard process was based on a namelist key (LEVELGPSRO), either 1 for
Bending, or 2 for refractivity. We wished to switch to a data-based
approach: LEVELGPSRO=3, process each profile according to the data that
are tagged active (either Bending or Refractivity within a given profile,
both within the same profile are impractical).
This allows a preprocessing that prepares a combination of both
approaches with different profiles.
Uses of LEVELGPSRO in the code were reduced strictly to drive the
filtering, but not the use of observation operators. These will be based
on the data they are acting upon (checking the variable type OBS_VNM).
The type of each profile is stored for simplicity during initial scan in
a previosuly existing store, IndexPrf, which was expanded to contain,
besides the variable number, other profile-based properties, such as
decoded satellite details and scanning mode (rising or setting in
altitude).
For normal RO files, which are homogeneous across all profiles, there
is no impact by this change. Timing and memory impact are negligible.