Issue #609: Using YbTinvRYb_mod in CVLETKF-ME algorithm.
YbTinvRYb_mod is used for computing perturbation update in
CVLETKF-ME algorithm.

memberIndexSubEns_mod and memberIndexSubEnsComp_mod arrays store
the modulated ensemble indices.

YbTinvRYb_CV_mod is equivalent to YbTinvRYb_CV for modulated
ensembles. eigenValues_CV_mod and eigenVectors_CV_mod are computed
from YbTinvRYb_CV_mod and used for computing the weight matrix with
CV.

The first identity matrix (Eq. 1 of Mark's LETKF paper) cannot be
included in the weight matrix and removed. This is because the weight
matrix is applied to the modulated ensembles to obtain the updates
for the original ensembles.