Issue #609: Correct computing final member pert.
Because identity matrix is removed from calculation of weightsMembers
in 75d5a77a, the resulting product is not weight but increment. The
increment should be added to (instead of subtracted from) background
ensemble perturbation to get member analysis perturbation.

When removing the weights mean from weightsMembersLatLon, the mean is
computed over nEns (which is the dimension of the array).