Issue #654: Avoid using 'printf' to convert spaces into newlines
We were using a combination of 'printf' and 'sed' to a serie of files
on a single line to a list containing one file per line.  But this
command was aborting when we reach the maximum line length given to
'printf'.

We avoid this by using "tr ' ' '\n'" which is a tool exactly for this.