Issue #658: missing lat/lon in burpread; sqldiff in check.tsk
There were remaining data entries with zero lat/lon for the added
data entries for wind and transformed variables.
Those are now initialized with missing value.

Adding new empty columns in diag sqlite files caused check.tsk
to flood the listings with the entire content of all the new and old sqlite
tables.

This commit proposes a solution with the sqldiff binary found at:
https://www.sqlite.org/download.html

sqldiff should probably be provided through an SSM somewhere? I did not include
it in this commit.