Issue #626: New variable considerSHIPstnID in thinning_mod.\n 20001(visibility), 50001(log visibility) and 11041(gust) are added in thn_surfaceInTime.
The original surface thinning process considers all 'stnids = SHIP' reports as being from a single station, because stnids is considered as a condition. The modified code introduced a new variable in the namelist and in ftn code: considerSHIPstnID. When considerSHIPstnID = .false., the stnids will not be used as a criteria for judging if the reports are from same station. After the modification, more 'SHIP' reports are preserved.

The original code does not keep 20001 (visibility) and 11041 (gust) as selected obs because they are not assimilated operationally. The new code adds those two variables, as well as an internal variable log(vis) 50001. After the modification, if and only if input observation file contains 20001 and 11041, the thinning process can keep them. However this modification has no impact in operation because the input doesn't include those two variables. Currently this modification impacts only 2d surface assimilation where 20001 and 11041 exist in derialt.