Issue #677: New subroutine ocebg_bgCheckSeaIce
This is where the new subroutine ocebg_bgCheckSeaIce
for the gross error check is introduced. A rms difference
between the background and the observations (OmP) over an
ice chart or a satellite swath is calculated. If the RMSD
is larger than a threshold provided in the namelist for that
observation type, then the entire ice chart or satellite swath
is rejected.