Issue #677: Remove bit 0 from the list of flags
This is because bit 0 is set for the regional ice charts where
the background-error standard deviation is small, indicating
that observations have been assimilated recently. The regional
ice charts are assimilated only where the background has not
seen observations for a long time. However, in the context of
obsSelection, the full chart needs to be considered such that
the RMSD between the background and the ice chart is meaningful
for the gross error check.