Issue #718: New file bgSeaIceConc
For the assimilation of ASCAT data for sea ice,
the ice concentration is needed in the observation-error
standard deviation calculation.
That means that for the analysis-error calculations,
both the background-error std dev. and the ice concentration itself
are required. So having the ice concentration field in a separate file
(bgSeaIceConc) is a solution in this context, to avoid a clash between
the field and its error, since both have the same nomvar and only
differ by their etiket.