Issue #720: Update of the GDPS unit test
In the end, they was no bug in the ozone-related code, simply a mismatch in the B file for ozone that needed to be updated to be compatablie with the latest version of the code.

The new unit test is fully fonctionnal and permit to test MIDAS against the GDPS OPS config.