Issue #733: ensembleH can read subset of members that do not start with 1.
fileMemberIndex1 is first member number to read from ensembles and is
the new variable to NAMENSEMBLEH namelist. The value is given to
ens_allocate and eob_allocate and stored as new component of the derived
types.

The value is available to the subroutines that do the reading and writing
of files and used in the code that determines the file names to be read/written.