Issue #733: Bug fix for assigning memberIndex of subset of member to ensObsGain.
A new subroutine eob_getMemebrIndexInFullEnsSet is created to return the
correct memberIndex corresponding to the full ensemble set. The
memberIndex array is used as input to the eob_writeToFilesMpiLocal and
eob_writeToFilesMpiGlobal subroutines to store correct memberIndex in
the header of binary files.

Adding the numMembers and memberIndex to the header of mpi global binary files.