Issue #739: Removing the useRttovScatt logic.
runObsOperatorWithHydrometeors* logical variable is used to indicate
running with rttovScatt.

This means instead of prescribing rttovScatt for each instrument/platform
pair, the namelist variable instrumentNamesUsingHydrometeors determines if
the in the BT simulation for an instrument one should include hydrometeors
profiles (i.e. rttovScatt). All (or none) MHS will run in all-sky mode with
this change if MHS is in instrumentNamesUsingHydrometeors array.

Bugfix for when running clear-sky RTTOV. Cloud profiles were not properly
stored/restored.