Issue #739: Match BURP and `obsSpaceData` in `brpr_addCloudParametersandEmissivity`.
To ensure match between the burp file reports and obsSpaceData when
updating the cloud variables, first we loop thought the reports
sequentially and store the reportIndex where there is any observation in
`ObsSpaceData`. We identify the first headerIndex. Then we start the update
loop on reports only on the ones where there is observation in
`ObsSpaceData`.