Issue #740: Implementation ofan ensemble mode for the oMinusF program
In this implemention, the approach to compute Hx in ensembleH was copied but the outputs are written in a SQL file using the ensobs option in sqlr_writeSqlDiagFile from sqliteread_mod

OminusF in ensemble mode therefore outputs Hx instead of y - Hx in the deterministc mode. However, the y value is available in the same obs file making the computation of y - Hx quite easy in a post-processing/verification steps.