Issue #848: Update to HFSO_dryNorm and HFSO_wetNorm test results
The difference is due to the subroutine obsu_setGbgpsError which sets the initial value of obs_oer for all GB-GPS observations with element ID bufr_nezd (which is the "zenith total delay" or ZTD measured by the ground-based GPS instrument). This subroutine is called very early, just after the observation file is read (either in burpFiles_mod or sqliteFiles_mod). So, due to the change made to oer_fillObsErrors where OBS_OER is set to missing if it is flagged to NOT be assimilated (i.e. in the final else near the end of the subroutine), the value of OBS_OER is now missing for rejected GB-GPS observations instead of equal to the "formal error value" as set in obsu_setGbgpsError (which comes from OBS_VAR, but for the element ID corresponding to the "formal error" which is bufr_nefe, and not from the actual observation which has element id bufr_nezd).