Add the version information in the email sent by 'maestro/suites/midas_system_tests/extractRunTime.sh'
When the system tests finishes in the CI, we send an email informing
the maintainers about the outliers in the time execution.

This information was not very useful on hand because the email did not
contain any information about the MIDAS version that has been tested.

Now, we include that information in the email.