Issue #898: Read flnml from string in program 'var1D' and modules 'calcStatsGlb_mod', 'simulateEmissivity_mod'
In #886, all namelist reads have been converted to read from a string
instead of the file itself.  This avoids some slowness when running
with many many MPI ranks.  This has been done in the branch
'v_3.9-RandD'.  We complete the work in places in the code from the
'main' branch that was not in 'v_3.9-RandD'.