Issue #850: Fix issue with gz2alt conversion for diagnostic levels
In GEM-P mode, we don't apply the gz2alt conversion for the diagnostic levels. Because this conversion was applied in GEM-H mode, this was introducing some inconsistencies in the values used for the vertical interpolation. To preserve the results obtained in GEM-P mode, the gz2alt conversion was removed for the diagnostic levels in GEM-H mode.

Also, the routine czp_ensureCompatibleTops was modified to be able to deal with a 21001 (GEM-H) grid, but only for interpolation where the source and the destination is on a 21001 grid.