Issue #879: Inflating obs error in all-sky is now instrument-specific.
`instrumentNamesInflateErrAllskyTt` and `instrumentNamesInflateErrAllskyHu` are
added to `namoer` namelist block and are used to prescribe the name of instruments
to inflate error in all-sky.

This is needed because error inflation is not used for MHS during IC-4, but could
be needed for ATMS humidity channels. Setting the `mwAllskyHuInflateByOmp` or
`mwAllskyHuInflateBySiDiffUsing` without this functionality will trigger error inflation
for all-sky humidity channels of all instruments.