Issue #914: New NAMINC option spreadIceIncOverLakes
The default behavior is spreadIceIncOverLakes = false,
i.e. the sea ice is clipped at zero and one.
When set to true, the lake sea ice excess to one or the
deficit from zero is spread over the lake,
except for the Great Lakes.
Elsewhere, the sea ice is simply clipped at zero and one.