Issue #932: Add 'LETKF' namelist variable 'maxGridSpacingAllowed'
A limit of 1000km was hard coded as the maximum grid spacing.  This is
a check to catch situations where we have trouble with computing the
grid spacing for a particular type of grid (this was a challenge for
the ORCA grid used for ocean and sea ice).

With a default of 1000 km, we can use that variable to change that
limit dynamically.

We set 'maxGridSpacingAllowed' to '1.2d6' for the test 'Tests/letkf/glb_800km'.