Issue #932: Add a 'NAMHCO' namelist variable 'maxGridSpacingAllowed'
A check was made in 'hco_SetupFromFile' to avoid a grid spacing of
more than 1000km.  This is a check to catch situations where we have
trouble with computing the grid spacing for a particular type of grid
(this was a challenge for the ORCA grid used for ocean and sea ice).

This threshold was hard-coded.

Here we introduced a new namelist block 'NAMHCO' which is read by the
routine 'hco_SetupFromFile' from the module 'horizontalCoord_mod' to
change that limit dynamically with the namelist variable
'maxGridSpacingAllowed' which has a default of 1000 km which was the
previous threshold.

We set 'maxGridSpacingAllowed' to '1.2d6' for the test
'Tests/letkf/glb_800km'.