Issue #922: Use parallel IO in 'gio_writeToFile' only when using "outputFormat='RSF'"
The code doing parallel IO is not there yet but we have prepare the
ground work for doing that.

The reference outputs of the test '/Tests/letkf/glb_15km' have been
converted to RSF to be able to compare them with the outputs directly
in RSF using the script 'convert2RSF.sh' in the reference directory.