Issue #934: Allocate the minimum memory possible for 'gd_send_r4' in 'ens_writeEnsemble'
Now that we use 'mpi_alltoallv', we send only the data that is really
needed.  So we can be more precise in the memory usage we do.  In
particular, the array 'gd_send_r4' can be allocated for that last
dimension with 'min(ens%numMembers, mmpi_nprocs)' since the data is
sent to assemble all the members on the first
'min(ens%numMembers, mmpi_nprocs)' MPI procs.