Issue #943: The 'theadId' was not set properly in the 'varLevIndex' loop in 'gio_writeToFile'.
We were taking into account the offset of 'varLevIndex' since we
switch the loop over 'varLev's instead of 'lev's.  So, if the loop
started with 'varLevIndex=50' and 'numThreadsForWriting=2', then
   threadId = mod(varLevIndex-1, numThreadsForWriting)

and 'threadId=1' which is 'numThreadsForWriting-1' and so the writing
to file is triggered for all threads from 0 to 1 but the thread is 0.