Adding argument to 'set_resources_def.sh' to specify the toplevel directory
During the 'install' step of the compilation process,
'set_resources_def.sh' is called from the build directory which may be
outside of the code repository.  So when grabbing the toplevel
directory with the Git command 'git rev-parse --show-toplevel', it
cannot since it is outside of the source code directory.

So we add an optional argument to 'set_resources_def.sh' which is the
top level source code directory.  The calling script can then give
this information to 'set_resources_def.sh'.

For all the calls of 'set_resources_def.sh', we add that new argument.