Issue #995: Rename the test '/Tests/var/EnVar/rdps' to '/Tests/var/EnVar/hrdps'
We use the operational configuration ran on 2024112906.  Using version
'3.9.1'.

So we start with a version of reference files at '0000' since it is
based on version '3.9.3' which valids with the operational
configuration.

We will upgrade the configuration and results when we will run that
test in the context of 'main' branch.