Issue #905: Changes to calcHeightAndPressure_mod
The needed changes in the above module and a few others were made to adapt them to an analysisGrid on vCode=21001 (i.e. height-based coordinate)

The program var.abs is now "working" with an analysisGrid on vCode=21001 but the minimization diverge due an issue in the adjoint version of the operator to compute pressure increments.