Issue #867: Introduced default values for  &THIN_TOVS namelist parameters. Removed write statements used during code development.
The substantive changes are mostly in subroutines thn_tovsFilt() and  thn_tovsfilt_dd().

The default thinning technique is set to grid-based. The default value for
the mindist parameter is set to a negative value to make sure that the user
sets this parameter to a positive value in the namelist.

delta and deltrad are checked for positivity.

The reading of namelist thin_tovs_dd is deleted in thn_tovsfilt_dd(). This is because
the parameters for both types of thinning are included in the same namelist
&THIN_TOVS.